Permission fixes UI fix

+++ resources/views/driver/show.blade.php
@@ -57,6 +57,7 @@
                                         {{$driver->phone}}
                                     </div>
                                     <div class="col-7 text-right  pt-2">
+                                        @if (Auth::user()->role_id > 3)
                                             @component('layouts.components.modal',[
                                                 'modelName'=>'driver',
                                                 'action'=>'edit',
@@ -68,6 +69,7 @@
                                                 'file'=>true
                                             ])
                                             @endcomponent
+                                        @endif
                                     </div>
                                 </div>
                             </div>
@@ -83,6 +85,13 @@
                     <table class="table">
                         <thead>
                             <tr>
+                                <td colspan="6">
+                                    <div class="pagination">
+                                        {{-- {{ $rides->links("pagination::bootstrap-4") }} --}}
+                                    </div>
+                                </td>
+                            </tr>
+                            <tr>
                                 <th>ID</th>
                                 <th>For</th>
                                 <th>{{--($op1=='Client') ? 'Invoice': 'Part'--}}#</th>
@@ -93,13 +102,13 @@
                             </tr>
                         </thead>
                         <tbody>
-                            @foreach ($driver->rides as $key => $ride)
+                            @foreach ($rides as $key => $ride)
                                 <tr>
                                     <td>{{$ride->id}}</td>
                                     <td>
                                         @component('layouts.components.tooltip',['modelName'=>'location','model'=>$ride->rideable->location])@endcomponent
                                     </td>
-                                    <td>
+                                    <td class="fixedWidthFont">
                                         @component('layouts.components.tooltip',['modelName'=>'rideable','model'=>$ride->rideable])@endcomponent
                                     </td>
                                     <td>
@@ -110,6 +119,7 @@
                                     </td>
                                     <td><span title="{{$ride->created_at}}">{{$ride->created_at->diffForHumans()}}</span></td>
                                     <td>
+                                        @if (Auth::user()->role_id > 3)
                                         @component('layouts.components.modal',[
                                             'modelName'=>'ride',
                                             'action'=>'edit',
@@ -120,11 +130,18 @@
                                             'op2'=>''
                                         ])
                                         @endcomponent
-                                        <a class="badge badge-danger" href="/ride/delete/{{$ride->id}}"> Delete</a>
-
+                                            <a class="badge badge-danger" href="/ride/delete/{{$ride->id}}"> Delete</a>
+                                        @endif
                                     </td>
                                 </tr>
                             @endforeach
+                            <tr>
+                                <td colspan="6">
+                                    <div class="pagination">
+                                        {{-- {{ $driver->rides->links("pagination::bootstrap-4") }} --}}
+                                    </div>
+                                </td>
+                            </tr>
                         </tbody>
                     </table>
                 </div>
@@ -162,18 +179,20 @@
                                     <td>{{$fillup->mileage}}</td>
                                     <td><span title="{{$fillup->created_at}}">{{$fillup->created_at->diffForHumans()}}</span></td>
                                     <td>
-                                        @component('layouts.components.modal',[
-                                            'modelName'=>'fillup',
-                                            'action'=>'edit',
-                                            'iterator'=>$key,
-                                            'object'=>$fillup,
-                                            'btnSize'=>'small',
-                                            'op1'=>'',
-                                            'op2'=>'',
-                                            'file'=>true
-                                        ])
-                                        @endcomponent
-                                        <a class="badge badge-danger" href="/fillup/delete/{{$fillup->id}}"> Delete</a>
+                                        @if (Auth::user()->role_id > 3)
+                                            @component('layouts.components.modal',[
+                                                'modelName'=>'fillup',
+                                                'action'=>'edit',
+                                                'iterator'=>$key,
+                                                'object'=>$fillup,
+                                                'btnSize'=>'small',
+                                                'op1'=>'',
+                                                'op2'=>'',
+                                                'file'=>true
+                                            ])
+                                            @endcomponent
+                                            <a class="badge badge-danger" href="/fillup/delete/{{$fillup->id}}"> Delete</a>
+                                        @endif
                                     </td>
                                 </tr>
                             @endforeach
